Factorize code and resolve bug in dqn_ann
